+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.0.58] - 2021-10-12
 ### Added
 - Added provider for APIKey
+- Configure the type of TokenProvider: header vs custom
 
 ## [0.0.54] - 2021-10-10
 ### Added